Fix variant-limiting bug in matdbg (#2903)

+++ libs/matdbg/include/matdbg/JsonWriter.h
@@ -45,7 +45,7 @@ public:
     // shader index is an active variant. Each bit in the activeVariants bitmask
     // represents one of the possible variant combinations.
     bool writeActiveInfo(const filaflat::ChunkContainer& package, backend::Backend backend,
-            uint16_t activeVariants);
+            uint64_t activeVariants);
 
 private:
     utils::CString mJsonString;
